requirements.txt
pytube>=15.0.0
openai-whisper @ git+https://github.com/openai/whisper.git
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
ffmpeg-python>=0.2.0
google-generativeai>=0.3.0
flask==3.0.0
ipython>=8.0.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
flask>=2.0.0
ipython>=8.0.0 